fix(pack-up): ensure we pass the tsconfig file name, not a path to it
fix(admin): annotate return type on createBuildContext

fix(admin): ensure shared is a the base folder not importing from server

chore(pack-up): PRINT MOREEEE

fix(ctb): export required type

chore: refactor tsconfigs
