+++ packages/core/admin/admin/tsconfig.build.json
@@ -1,9 +1,23 @@
 {
-  "extends": "./admin/tsconfig.json",
-  "include": ["./admin/src/*", "./admin/custom.d.ts", "./shared/*", "./admin/module.d.ts"],
-  "exclude": ["tests/*", "**/*.test.*"],
+  "extends": "tsconfig/client.json",
   "compilerOptions": {
-    "rootDir": ".",
+    "rootDir": "../",
+    "baseUrl": ".",
     "outDir": "./dist"
-  }
+  },
+  "include": [
+    "./src",
+    "./custom.d.ts",
+    "../shared",
+    "./module.d.ts",
+    "../ee/admin",
+    "../package.json"
+  ],
+  "exclude": [
+    "./tests",
+    "**/__mocks__",
+    "**/*.test.*",
+    "../ee/admin/**/*.test.*",
+    "../ee/admin/**/__mocks__"
+  ]
 }
