--- packages/core/admin/ee/admin/tsconfig.build.json
@@ -1,7 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "dist"
-  },
-  "exclude": ["tests", "**/*.test.*"]
-}
