+++ packages/core/admin/ee/admin/tsconfig.json
@@ -1,11 +1,11 @@
 {
   "extends": "tsconfig/client.json",
   "compilerOptions": {
+    "rootDir": "../",
     "baseUrl": ".",
     "paths": {
       "@tests/*": ["../../admin/tests/*"]
     }
   },
-  "include": ["src", "tests", "custom.d.ts", "module.d.ts"],
-  "exclude": ["node_modules"]
+  "include": ["src", "tests", "custom.d.ts"]
 }
