+++ packages/core/admin/ee/server/tsconfig.json
@@ -1,9 +1,10 @@
 {
   "extends": "tsconfig/base.json",
   "include": ["src"],
-  "exclude": ["node_modules"],
   "compilerOptions": {
     "types": ["lodash", "jest"],
+    "rootDir": "../../",
+    "baseUrl": ".",
     "esModuleInterop": true,
     "allowJs": true
   }
