--- packages/core/admin/server/src/tsconfig.build.json
@@ -1,9 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "include": ["./src"],
-  "exclude": ["./src/**/*.test.ts"],
-  "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "./dist/server"
-  }
-}
