+++ packages/core/admin/server/tsconfig.json
@@ -1,8 +1,9 @@
 {
   "extends": "tsconfig/base.json",
   "include": ["src"],
-  "exclude": ["node_modules"],
   "compilerOptions": {
+    "rootDir": "../",
+    "baseUrl": ".",
     "types": ["lodash", "jest"],
     "esModuleInterop": true
   }
