+++ packages/core/admin/tsconfig.json
@@ -1,5 +1,7 @@
 {
   "extends": "tsconfig/client.json",
   "include": ["_internal", "packup.config.ts"],
-  "exclude": ["node_modules"]
+  "compilerOptions": {
+    "outDir": "./dist/cli"
+  }
 }
