+++ packages/core/content-type-builder/admin/tsconfig.json
@@ -1,5 +1,8 @@
 {
   "extends": "tsconfig/client.json",
   "include": ["src"],
-  "exclude": ["node_modules"]
+  "compilerOptions": {
+    "rootDir": "../",
+    "baseUrl": "."
+  }
 }
