+++ packages/core/content-type-builder/server/tsconfig.json
@@ -1,9 +1,10 @@
 {
   "extends": "tsconfig/base.json",
   "include": ["src"],
-  "exclude": ["node_modules"],
   "compilerOptions": {
     "types": ["lodash", "jest"],
-    "esModuleInterop": true
+    "esModuleInterop": true,
+    "rootDir": "../",
+    "baseUrl": "."
   }
 }
