+++ packages/core/email/admin/tsconfig.build.json
@@ -1,9 +1,10 @@
 {
-  "extends": "./admin/tsconfig.json",
-  "include": ["./admin/src", "./shared"],
+  "extends": "./tsconfig.json",
+  "include": ["./src", "../shared"],
   "exclude": ["**/*.test.ts"],
   "compilerOptions": {
-    "rootDir": ".",
+    "rootDir": "../",
+    "baseUrl": ".",
     "outDir": "./dist"
   }
 }
