+++ packages/core/email/server/tsconfig.build.json
@@ -1,9 +1,10 @@
 {
   "extends": "./server/tsconfig.json",
-  "include": ["./server/src"],
+  "include": ["./src", "../shared"],
   "exclude": ["./server/src/**/*.test.ts"],
   "compilerOptions": {
-    "rootDir": ".",
+    "rootDir": "../",
+    "baseUrl": ".",
     "outDir": "./dist"
   }
 }
