+++ packages/core/email/server/tsconfig.json
@@ -1,8 +1,9 @@
 {
   "extends": "tsconfig/base.json",
-  "include": ["src"],
-  "exclude": ["node_modules"],
+  "include": ["./src", "../shared"],
   "compilerOptions": {
+    "rootDir": "../",
+    "baseUrl": ".",
     "esModuleInterop": true
   }
 }
