+++ packages/plugins/cloud/admin/tsconfig.json
@@ -1,5 +1,8 @@
 {
   "extends": "tsconfig/client.json",
-  "include": ["src", "global.d.ts", "module.d.ts"],
-  "exclude": ["node_modules"]
+  "include": ["./src", "*.d.ts"],
+  "compilerOptions": {
+    "rootDir": "../",
+    "baseUrl": "."
+  }
 }
