+++ packages/plugins/color-picker/server/tsconfig.build.json
@@ -1,8 +1,10 @@
 {
-  "extends": "./server/tsconfig.json",
-  "include": ["./server/src"],
+  "extends": "./tsconfig.json",
+  "include": ["./src"],
   "exclude": ["**/*.test.ts"],
   "compilerOptions": {
-    "outDir": "./dist/server"
+    "rootDir": "../",
+    "baseUrl": ".",
+    "outDir": "./dist"
   }
 }
