+++ packages/plugins/i18n/admin/tsconfig.build.json
@@ -1,9 +1,10 @@
 {
-  "extends": "./admin/tsconfig.json",
-  "include": ["./admin/src", "./shared"],
+  "extends": "./tsconfig.json",
+  "include": ["./src", "../shared"],
   "exclude": ["tests", "**/*.test.*"],
   "compilerOptions": {
-    "rootDir": ".",
+    "rootDir": "../",
+    "baseUrl": ".",
     "outDir": "./dist"
   }
 }
