+++ packages/plugins/i18n/server/tsconfig.json
@@ -1,5 +1,8 @@
 {
   "extends": "tsconfig/base.json",
   "include": ["./src"],
-  "exclude": ["node_modules"]
+  "compilerOptions": {
+    "rootDir": "../",
+    "baseUrl": "."
+  }
 }
