+++ packages/utils/pack-up/tsconfig.build.json
@@ -5,5 +5,5 @@
     "outDir": "./dist"
   },
   "include": ["src"],
-  "exclude": ["tests", "**/__tests__/**", "**/cli/**", "./src/types.ts"]
+  "exclude": ["tests", "**/__tests__/**", "**/cli/**"]
 }
